Split joiner_test tests into separate files
